Fix servo speed/threading glitch
The servos (and possibly motors too) have been glitching out when the speed is set to 1, because the robot tries to sleep for 0 milliseconds in between each tick. This caused us some issues with sleeping in other parts of the program and made the servos unreliable. This commit should fix those issues.